ex 4.22 + tests before token auth. (4.12)

+++ part4.bloglist/controllers/blogs.js
@@ -1,16 +1,33 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
-blogsRouter.get('/', (request, response) => {
-  Blog
-    .find({})
-    .then(blogs => {
-      response.json(blogs)
-    })
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
+  response.json(blogs)
 })
 
-blogsRouter.post('/', (request, response) => {
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog) {
+    response.json(blog)
+  } else {
+    response.status(404).end()
+  }
+})
+
+blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
+  const user = request.user
+
+  if (!user.id) {    
+    return response.status(401).json({ 
+      error: 'token missing or invalid' 
+    })  
+  }  
+
 
   if (body.title === undefined) {
     return response.status(400).json({error: 'title missing'})
@@ -20,35 +37,53 @@ blogsRouter.post('/', (request, response) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes || 0,
+    user: user.id
   })
 
-  blog
-    .save()
-    .then(savedBlog => {
-      response.json(savedBlog)
-    })
-    .catch(error => next(error))
+  const savedBlog = await blog.save()
+
+  const userToUpdate = await User.findById(user.id)
+
+  userToUpdate.blogs = userToUpdate.blogs.concat(savedBlog._id)
+  await userToUpdate.save()
+
+  response.json(savedBlog)
 })
 
-blogsRouter.get('/:id', (request, response, next) => {
-  Blog.findById(request.params.id)
-  .then(blog => {
-    if (blog) {
-      response.json(blog)
-    } else {
-      response.status(404).end()
-    }
-  })
-  .catch(error=>next(error))
+
+blogsRouter.delete('/:id', async (request, response) => {
+  
+  blogToDelete = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (!user.id) {    
+    return response.status(401).json({ 
+      error: 'token missing or invalid' 
+    })  
+  }  
+
+   if ( blogToDelete.user.toString() === user.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+   } else {
+    return response.status(401).json({ 
+      error: "You don't have permission to delete this blog"
+    })  
+   }
+
 })
 
-blogsRouter.delete('/:id', (request, response, next) => {
-  Blog.findByIdAndRemove(request.params.id)
-    .then(result => {
-      console.log(request.params.id)
-      response.status(204).end()
-    })
-    .catch(error => next(error))
+blogsRouter.put('/:id', async (request, response) => {
+  const blog = {
+    title: request.body.title,
+    author: request.body.author,
+    url: request.body.url,
+    likes: request.body.likes
+  }
+
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new:true})
+  response.json(updatedBlog)
 })
+
 module.exports = blogsRouter
